Move run to its own file
