+++ scala/src/main/scala/rays/run.scala
@@ -1,0 +1,57 @@
+package rays
+
+import rays.bounds.AxisAlignedBox
+import rays.math.{Ray, Vector3}
+import rays.scenes.{Node, Scene}
+import rays.shapes.{Box, CollisionShape, Sphere, Triangle}
+
+import lcg.Random
+
+import java.lang.Long.remainderUnsigned
+
+/** The bounds of the world in which the simulation occurs. */
+val worldBounds = AxisAlignedBox(Vector3.zero, Vector3(100, 100, 100))
+
+/** Returns a random collision shape that fits in a unit box centered at the origin. */
+def randomCollisionShape(using g: Random): CollisionShape =
+  remainderUnsigned(g.nextLong(), 3) match
+    case 0 =>
+      Box.unit
+    case 1 =>
+      Sphere.unit
+    case 2 =>
+      Triangle(
+        Vector3.randomIn(AxisAlignedBox.unit),
+        Vector3.randomIn(AxisAlignedBox.unit),
+        Vector3.randomIn(AxisAlignedBox.unit))
+    case n =>
+      throw new RuntimeException("unreachable: " + n)
+
+def run(size: Int): Int =
+  given Random = Random(0xACE1)
+
+  // Create an empty world.
+  var world = Scene()
+
+  // Add random objects to the world.
+  for _ <- 0 until size do
+    val n = Node(Vector3.randomIn(worldBounds), Some(randomCollisionShape))
+    world = world.adding(n)
+
+  // Shoot rays at all objects from the origin to find those that are (partially) occluded.
+  var nodes = Set.from(world.nodes)
+  var occluded = List[Node]()
+
+  for n <- world.nodes if nodes(n) do
+    //println(s"Node: ${n.translation}")
+    val s = n.shape.get
+    val r = Ray(Vector3.zero, (world.translation(n) + s.centroid).normalized)
+    //println(s"Ray: $r")
+    val collisions = world.shoot(r, -1, false)
+    for (m, _) <- collisions do
+      n.collisionMask = 0
+      nodes -= m
+    //println(collisions)
+    occluded ++= collisions.drop(1).map((c) => c(0))
+
+  occluded.length
